data_loading/pyproject.toml
[project]
name = "chess-graph"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "neo4j",
    "python-chess",
    "python-dotenv",
    "httpx",
]

[project.scripts]
chess-graph = "chess_graph.__main__:main"
sample-queries = "chess_graph.sample_queries:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chess_graph"]
